Added travis-CI script to deploy doxygen-documentation to github pages ( https://openmw.github.io )
The documentation can be updated by merging into the doxygen branch, this will trigger a new travis build and documentation update.

GH_REPO_TOKEN must be set up in travis-ci settings, this is a github API token for a minimally priviledged github user with access to the openmw.github.io repo.